fix: work around configure issue with GCC 14+
See https://github.com/owasp-modsecurity/ModSecurity-nginx/pull/275.
ModSecurity-nginx uses `printf` as part of the nginx configure test. As
of GCC 14, this leads to an error during compilation and thus to an
error when running `configure` on nginx.

This commit works around the issue by patching the `config` file in
ModSecurity-nginx with a valid value for.